# Refactoring of the parsing of simple units, moving short-circuit up