--- a/docs/changes/units/17012.perf.rst
+++ b/docs/changes/units/17012.perf.rst
@@ -1,0 +1,1 @@
+Sped up the parsing of simple units by about 25%.