This is an alternative to #17004 that does not rely on private methods, and removes the short-circuiting from the unit parser. It builds on #17011, since removing the short-circuit from the unit parser brought out quite a few bugs. (EDIT: #17011 is merged, so only 3 commits left.) But it turns out that for FITS at least, some short-circuiting has to be done, since it weirdly defines a unit "deg C", which would in normal parsing become "degrees Coulomb" (and which we really do *not* want in a composite unit!).

Note that while the last commit is labelled performance, this is only relative to the case where the short circuit is not present at all. Relative to main, performance benefits are modest.

- [X] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.