@mhvk 2024-09-15
Benchmarks are a bit weird: simple units are faster, as expected, but quantity initialization has slowed down? Will try to rerun, especially given the large variance, but for the record:
```
| Change   | Before [b266a404]    | After [b02b73a0]    |   Ratio | Benchmark (Parameter)             |
|----------|----------------------|---------------------|---------|-----------------------------------|
| -        | 2.46±0.01μs          | 1.85±0.01μs         |    0.75 | units.time_very_simple_unit_parse |
| -        | 123±70μs             | 48.6±0.2μs          |    0.39 | units.time_quantity_times_unit    |
| Change   | Before [b266a404]    | After [b02b73a0]    |   Ratio | Benchmark (Parameter)                |
|----------|----------------------|---------------------|---------|--------------------------------------|
| +        | 38.6±0.2μs           | 126±90μs            |    3.25 | units.time_quantity_creation         |
| +        | 97.9±1μs             | 234±100μs           |    2.39 | units.time_quantity_init_array       |
| +        | 168±1μs              | 309±100μs           |    1.84 | units.time_quantity_array_conversion |
``