@mhvk 2024-09-30
Hmm, looking at your PR again, I think it is really equally simple (after the other changes made), and does cover the FITS case. So, let me close this one in favour of #17004.